# Optimize slice decoding