I tested this package against the easyjson benchmarks and there was one benchmark where we were significantly slower. After digging further it appeared to be related to calling `reflect.MakeSlice`, which forces an intermediary heap allocation to store the slice value in a `reflect.Value`. Decoders that use code generation can call `make` to construct slices and avoid this cost.

The first commit in this PR shows how to reduce the number of heap allocations while still using the `reflect.` package (and therefore guaranteeing forward compatibility), the second commit makes a direct call to `unsafe_NewArray`. Here are the diff for each code change compared to master:
```
name                          old time/op    new time/op    delta
EasyjsonUnmarshalSmallStruct    1.11µs ± 2%    1.19µs ± 2%   +6.78%  (p=0.000 n=10+10)

name                          old alloc/op   new alloc/op   delta
EasyjsonUnmarshalSmallStruct      656B ± 0%      592B ± 0%   -9.76%  (p=0.000 n=10+10)

name                          old allocs/op  new allocs/op  delta
EasyjsonUnmarshalSmallStruct      6.00 ± 0%      4.00 ± 0%  -33.33%  (p=0.000 n=10+10)
```
```
name                          old time/op    new time/op    delta
EasyjsonUnmarshalSmallStruct    1.11µs ± 2%    0.91µs ± 2%  -17.87%  (p=0.000 n=10+10)

name                          old alloc/op   new alloc/op   delta
EasyjsonUnmarshalSmallStruct      656B ± 0%      592B ± 0%   -9.76%  (p=0.000 n=10+10)

name                          old allocs/op  new allocs/op  delta
EasyjsonUnmarshalSmallStruct      6.00 ± 0%      4.00 ± 0%  -33.33%  (p=0.000 n=10+10)
```